+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-# [Unreleased-main] - 2026-02-25
+# [Unreleased-main] - 2026-03-10
 
 ## Added
 - Parsing of ensemble profiles when using input profiles from csv.
 - Maximum profile constraint for PV asset is considered in PV sizing
+- Cater for input via esdl constraints to specify the upper limit for OPTIONAL assets in DTK
+- Initial implementation of adaptable pipe DN lower limit per pipe
 - heat_loss_efficiency of HeatStorage is read from dischargeEfficiency attribute in esdl